Move stdlib DataProducts and DataViews to companion objects
This makes them available by default in the implicit scope.
